% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortSqlBuilder.R
\name{buildCohortQuery}
\alias{buildCohortQuery}
\title{Build cohort SQL}
\usage{
buildCohortQuery(expression, options)
}
\arguments{
\item{expression}{result of cohortExpressionFromJson containing the cohort expression.}

\item{options}{The options built using createGenerateOptions()}
}
\value{
A character vector containing the OHDSI Sql for the cohort expression
}
\description{
Generates the OMOP CDM Sql to generate the cohort expression
}
