% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptSetSqlBuilder.R
\name{buildConceptSetQuery}
\alias{buildConceptSetQuery}
\title{Build conceptset SQL}
\usage{
buildConceptSetQuery(conceptSetJSON)
}
\arguments{
\item{conceptSetJSON}{a string containing the JSON for the conceptset expression.}
}
\value{
A character vector containing the OHDSI Sql for the conceptset expression
}
\description{
Generates the OMOP CDM Sql to resolve the concept set expression
}
