% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortChecker.R
\name{checkCohort}
\alias{checkCohort}
\title{Check cohort}
\usage{
checkCohort(expression)
}
\arguments{
\item{expression}{result of cohortExpressionFromJson containing the cohort expression}
}
\value{
A list of warning messages
}
\description{
Checks a cohort expression for warnings
}
