% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{cohortDefinitionJson}
\alias{cohortDefinitionJson}
\title{An example cohort definition}
\format{
A character vector containing a JSON string
}
\usage{
data(cohortDefinitionJson)
}
\description{
An example cohort definition
}
\details{
An example cohort expression in JSON format
}
\keyword{datasets}
