% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortExpression.R
\name{cohortExpressionFromJson}
\alias{cohortExpressionFromJson}
\title{Render read JSON into a CohortExpression instance}
\usage{
cohortExpressionFromJson(expressionJson)
}
\arguments{
\item{expressionJson}{A character vector containing the cohort expression.}
}
\value{
A RJSIONIO list representing the JSON structure.
}
\description{
Reads a String (json) and deserializes it into a org.ohdsi.circe.cohortdefinition.CohortExpression
}
