% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintFriendly.R
\name{cohortPrintFriendly}
\alias{cohortPrintFriendly}
\title{Render cohort print-friendly}
\usage{
cohortPrintFriendly(expression)
}
\arguments{
\item{expression}{A character vector or result of cohortExpressionFromJson containing the cohort expression.}
}
\value{
A character vector containing the markdown.
}
\description{
Generates a print-friendly (human-readable) representation of the cohort definition. This can for example 
be used in a study protocol.
}
