% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptSetExpression.R
\name{conceptSetExpressionFromJson}
\alias{conceptSetExpressionFromJson}
\title{Render read JSON into a ConceptSetExpression instance}
\usage{
conceptSetExpressionFromJson(expressionJson)
}
\arguments{
\item{expressionJson}{A character vector containing the cohort expression.}
}
\value{
A java instance of org.ohdsi.circe.cohortdefinition.CohortExpression.
}
\description{
Reads a String (json) and deserializes it into a org.ohdsi.circe.vocabulary.CohortExpression
}
