% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{conceptSetJson}
\alias{conceptSetJson}
\title{An example concept set}
\format{
A character vector containing a JSON string
}
\usage{
data(conceptSetJson)
}
\description{
An example concept set
}
\details{
An example concept set in JSON format
}
\keyword{datasets}
