% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{conceptSetListJson}
\alias{conceptSetListJson}
\title{An example concept set list}
\format{
A character vector containing a JSON string
}
\usage{
data(conceptSetListJson)
}
\description{
An example concept set list
}
\details{
An example concept set list in JSON format
}
\keyword{datasets}
