% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintFriendly.R
\name{conceptSetListPrintFriendly}
\alias{conceptSetListPrintFriendly}
\title{Render conceptSet array for print-friendly}
\usage{
conceptSetListPrintFriendly(conceptSetList)
}
\arguments{
\item{conceptSetList}{A ConceptSet[] (from cohortExpression.conceptSets)}
}
\value{
A character vector containing the markdown.
}
\description{
Generates a print-friendly (human-readable) representation of an array of concept sets. This can for example 
be used in a study protocol.
}
