% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortSqlBuilder.R
\name{createGenerateOptions}
\alias{createGenerateOptions}
\title{Create Generation Options}
\usage{
createGenerateOptions(
  cohortIdFieldName,
  cohortId,
  cdmSchema,
  targetTable,
  resultSchema,
  vocabularySchema,
  generateStats
)
}
\arguments{
\item{cohortIdFieldName}{The field that contains the cohortId in the cohort table.}

\item{cohortId}{The generated cohort ID.}

\item{cdmSchema}{The value of the CDM schema.}

\item{targetTable}{the cohort table name.}

\item{resultSchema}{the schema the cohort table belongs to.}

\item{vocabularySchema}{the schema of the vocabulary tables (defaults to cdmSchema)}

\item{generateStats}{a boolean representing if the query should include inclusion rule statistics calculation}
}
\value{
A jObj that can be passed to generateSql().
}
\description{
Creates the generation options object for use in generateSql()
}
