% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Backend.R
\name{.get_backend}
\alias{.get_backend}
\title{Get the backend implementation}
\usage{
.get_backend()
}
\value{
An object with class 'circe_backend_python' or 'circe_backend_java'.
}
\description{
Get the backend implementation
}
\keyword{internal}
