% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Backend.R
\name{ensurePythonBackend}
\alias{ensurePythonBackend}
\title{Ensures the Python backend is available}
\usage{
ensurePythonBackend()
}
\value{
The 'circepy' module.
}
\description{
Ensures the Python backend is available
}
\keyword{internal}
