% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrettyPrint.R
\name{renderPrintFriendly}
\alias{renderPrintFriendly}
\title{Render print-friendly}
\usage{
renderPrintFriendly(expression)
}
\arguments{
\item{expression}{A character vector containing the cohort expression.}
}
\value{
A character vector containing the markdown.
}
\description{
Generates a print-friendly (human-readable) representation of the cohort definition. This can for example 
be used in a study protocol.
}
