% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Backend.R
\name{usePythonBackend}
\alias{usePythonBackend}
\title{Checks if the Python backend is enabled}
\usage{
usePythonBackend()
}
\value{
TRUE if CIRCER_BACKEND environment variable is set to 'python', FALSE otherwise.
}
\description{
Checks if the Python backend is enabled
}
