% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CirceR.R
\docType{package}
\name{CirceR-package}
\alias{CirceR}
\alias{CirceR-package}
\title{CirceR: Construct Cohort Inclusion and Restriction Criteria Expressions}
\description{
Wraps the 'CIRCE' (\url{https://github.com/ohdsi/circe-be}) 'Java' library allowing cohort definition expressions to be edited, and converted to 'Markdown' or 'SQL'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/CirceR/}
  \item \url{https://github.com/OHDSI/CirceR/}
  \item Report bugs at \url{https://github.com/OHDSI/CirceR/issues/}
}

}
\author{
\strong{Maintainer}: Chris Knoll \email{cknoll@ohdsi.org}

Authors:
\itemize{
  \item Martijn Schuemie
}

}
\keyword{internal}
